chnaged v method to work as sync instead of async
