+++ CHANGELOG.md
@@ -1,0 +1,13 @@
+
+# Change Log
+All notable changes to this project will be documented in this file.
+
+
+
+### Version 1.0.11
+You can call the method "getCustomersByBatchID" without the "await"
+since it is not an async operation.
+
+
+
+### Version 1.0.10
